requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-decouple>=3.8
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9
dj-database-url>=2.1